[Refactor] 네이밍 변경, 위치 변경, 불필요한 코드 주석
