+++ src/main/java/com/yzgeneration/evc/useditem/infrastructure/UsedItemEntity.java
@@ -40,9 +40,9 @@ public class UsedItemEntity {
 
     private int chattingCount;
 
-    @ElementCollection
-    @CollectionTable(name = "useditem_image", joinColumns = @JoinColumn(name = "used_items_id"))
-    private List<String> imageName = new ArrayList<>();
+//    @ElementCollection
+//    @CollectionTable(name = "useditem_image", joinColumns = @JoinColumn(name = "used_items_id"))
+//    private List<String> imageName = new ArrayList<>();
 
     private LocalDateTime createdAt;
 
@@ -57,7 +57,7 @@ public class UsedItemEntity {
                 .viewCount(usedItem.getViewCount())
                 .likeCount(usedItem.getLikeCount())
                 .chattingCount(usedItem.getChattingCount())
-                .imageName(usedItem.getImageName())
+//                .imageName(usedItem.getImageName())
                 .createdAt(usedItem.getCreatedAt()).build();
     }
 
@@ -73,7 +73,7 @@ public class UsedItemEntity {
                 .viewCount(viewCount)
                 .likeCount(likeCount)
                 .chattingCount(chattingCount)
-                .imageName(imageName)
+//                .imageName(imageName)
                 .createdAt(createdAt)
                 .build();
     }
