+++ src/main/java/com/yzgeneration/evc/verification/implement/EmailVerificationProcessor.java
@@ -2,23 +2,25 @@ package com.yzgeneration.evc.verification.implement;
 
 import com.yzgeneration.evc.common.service.port.UuidHolder;
 import com.yzgeneration.evc.verification.enums.EmailVerificationType;
-import com.yzgeneration.evc.verification.infrastructure.EmailRepository;
+import com.yzgeneration.evc.verification.infrastructure.EmailVerificationRepository;
 import com.yzgeneration.evc.verification.model.Email;
 import com.yzgeneration.evc.verification.model.EmailVerification;
-import com.yzgeneration.evc.common.service.port.MailSender;
+import com.yzgeneration.evc.member.service.port.MailSender;
+import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@Builder
 @RequiredArgsConstructor
 public class EmailVerificationProcessor {
 
-    private final EmailRepository emailRepository;
+    private final EmailVerificationRepository emailVerificationRepository;
     private final UuidHolder uuidHolder;
     private final MailSender mailSender;
 
     public EmailVerification createEmailVerification(Long userId, String emailAddress, EmailVerificationType emailVerificationType) {
-        return emailRepository.save(EmailVerification.create(userId, emailAddress, uuidHolder, emailVerificationType));
+        return emailVerificationRepository.save(EmailVerification.create(userId, emailAddress, uuidHolder, emailVerificationType));
     }
     
     public void sendMail(EmailVerification emailVerification) {
