+++ src/main/java/com/yzgeneration/evc/verification/infrastructure/EmailVerificationRepository.java
@@ -2,6 +2,6 @@ package com.yzgeneration.evc.verification.infrastructure;
 
 import com.yzgeneration.evc.verification.model.EmailVerification;
 
-public interface EmailRepository {
+public interface EmailVerificationRepository {
     EmailVerification save(EmailVerification emailVerification);
 }
